refactor: 버스 스케줄링 모듈 리팩토링
