+++ src/main/java/in/koreatech/koin/domain/bus/service/CityBusService.java
@@ -2,7 +2,9 @@ package in.koreatech.koin.domain.bus.service;
 
 import static in.koreatech.koin.domain.bus.model.enums.BusStation.getDirection;
 
+import java.time.Clock;
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
@@ -12,8 +14,13 @@ import java.util.Set;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import in.koreatech.koin.domain.bus.client.CityBusClient;
+import in.koreatech.koin.domain.bus.client.CityBusRouteClient;
 import in.koreatech.koin.domain.bus.dto.BusScheduleResponse;
+import in.koreatech.koin.domain.bus.dto.city.CityBusArrival;
+import in.koreatech.koin.domain.bus.dto.city.CityBusRoute;
 import in.koreatech.koin.domain.bus.dto.city.CityBusTimetableResponse;
+import in.koreatech.koin.domain.bus.exception.BusOpenApiException;
 import in.koreatech.koin.domain.bus.facade.route.CityBusRouteManager;
 import in.koreatech.koin.domain.bus.model.city.CityBusCache;
 import in.koreatech.koin.domain.bus.model.city.CityBusRemainTime;
@@ -21,10 +28,14 @@ import in.koreatech.koin.domain.bus.model.city.CityBusRouteCache;
 import in.koreatech.koin.domain.bus.model.city.CityBusTimetable;
 import in.koreatech.koin.domain.bus.model.enums.BusDirection;
 import in.koreatech.koin.domain.bus.model.enums.BusStation;
+import in.koreatech.koin.domain.bus.model.enums.BusStationNode;
 import in.koreatech.koin.domain.bus.model.enums.CityBusDirection;
 import in.koreatech.koin.domain.bus.repository.CityBusCacheRepository;
 import in.koreatech.koin.domain.bus.repository.CityBusRouteCacheRepository;
 import in.koreatech.koin.domain.bus.repository.CityBusTimetableRepository;
+import in.koreatech.koin.domain.version.model.VersionType;
+import in.koreatech.koin.domain.version.repository.VersionRepository;
+import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -37,6 +48,11 @@ public class CityBusService {
     private final CityBusRouteCacheRepository cityBusRouteCacheRepository;
     private final CityBusCacheRepository cityBusCacheRepository;
     private final CityBusTimetableRepository cityBusTimetableRepository;
+    private final CityBusClient cityBusClient;
+    private final CityBusRouteClient cityBusRouteClient;
+
+    private final VersionRepository versionRepository;
+    private final Clock clock;
 
     public List<CityBusRemainTime> getBusRemainTime(BusStation depart, BusStation arrival) {
         BusDirection direction = getDirection(depart, arrival);
@@ -90,5 +106,37 @@ public class CityBusService {
             .getByBusInfoNumberAndBusInfoArrival(busNumber, arrival.getName());
 
         return CityBusRouteManager.getCityBusSchedule(timetable, busNumber, depart, arrival, date);
+    }
+
+    @Transactional
+    public void storeRemainTime() {
+        List<List<CityBusArrival>> arrivalInfosList = new ArrayList<>();
+        BusStationNode.getNodeIds().forEach((nodeId) -> {
+            try {
+                arrivalInfosList.add(cityBusClient.getOpenApiResponse(nodeId).extractBusArrivalInfo());
+            } catch (BusOpenApiException ignored) {
+            }
+        });
+        LocalDateTime updatedAt = LocalDateTime.now(clock);
+        arrivalInfosList.forEach((arrivalInfos) -> cityBusCacheRepository.save(
+            CityBusCache.of(
+                arrivalInfos.get(0).nodeid(),
+                arrivalInfos.stream()
+                    .map(busArrivalInfo -> busArrivalInfo.toCityBusCacheInfo(updatedAt))
+                    .toList()
+            )
+        ));
+        versionRepository.getByType(VersionType.CITY).update(clock);
+    }
+
+    @Transactional
+    public void storeCityBusRoute() {
+        BusStationNode.getNodeIds().forEach((nodeId) -> {
+            try {
+                Set<CityBusRoute> routes = Set.copyOf(cityBusRouteClient.getOpenApiResponse(nodeId).extractBusRouteInfo());
+                cityBusRouteCacheRepository.save(CityBusRoute.toCityBusRouteCache(nodeId, routes));
+            } catch (BusOpenApiException ignored) {
+            }
+        });
     }
 }
