+++ src/main/java/in/koreatech/koin/domain/bus/util/BusScheduler.java
@@ -3,9 +3,7 @@ package in.koreatech.koin.domain.bus.util;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
-import in.koreatech.koin.domain.bus.client.CityBusClient;
-import in.koreatech.koin.domain.bus.client.CityBusRouteClient;
-import in.koreatech.koin.domain.bus.service.ExpressBusService;
+import in.koreatech.koin.domain.bus.facade.BusScheduleService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -14,15 +12,13 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class BusScheduler {
 
-    private final ExpressBusService expressBusService;
-    private final CityBusClient cityBusClient;
-    private final CityBusRouteClient cityBusRouteClient;
+    private final BusScheduleService busScheduleService;
 
     @Scheduled(cron = "0 * * * * *")
     public void cacheCityBusByOpenApi() {
         try {
-            cityBusClient.storeRemainTime();
-            cityBusRouteClient.storeCityBusRoute();
+            busScheduleService.storeCityBusRemainTime();
+            busScheduleService.storeCityBusRoute();
         } catch (Exception e) {
             log.warn("시내버스 스케줄링 과정에서 오류가 발생했습니다.");
         }
@@ -31,7 +27,7 @@ public class BusScheduler {
     @Scheduled(cron = "0 30 0 * * *")
     public void cacheExpressBusByOpenApi() {
         try {
-            expressBusService.storeRemainTimeByRatio();
+            busScheduleService.storeExpressBusRemainTime();
         } catch (Exception e) {
             log.warn("시외버스 스케줄링 과정에서 오류가 발생했습니다.");
         }
